move deployment class back to settings
