+++ backend/app/internal/flowsheet_manager.py
@@ -24,42 +24,13 @@ from pydantic import BaseModel, validator, field_validator, ValidationInfo, Fiel
 import tinydb  # JSON single-file 'database'
 
 # package-local
-from app.internal.settings import AppSettings
+from app.internal.settings import Deployment, AppSettings
 from watertap.ui.fsapi import FlowsheetInterface
 import idaes.logger as idaeslog
 
 _log = idaeslog.getLogger(__name__)
 _log.setLevel(idaeslog.DEBUG)
 VERSION = 3
-
-class Deployment:
-    """Values related to the deployment context of the UI,
-    e.g., NAWI WaterTAP, PROMMIS, or IDAES.
-    """
-    # env var for project name
-    PROJECT_ENV = "PSE_PROJECT"
-    # projects and their associated packages
-    PROJ = {
-        "nawi": ("watertap",),
-        "idaes": ("idaes",),
-        "prommis": ("prommis",)
-    }
-    DEFAULT_PROJ = "nawi"
-
-    def __init__(self, project=DEFAULT_PROJ):
-        _log.info(f"Deploy for project={project}")
-        if project not in self.PROJ.keys():
-            valid_projects = ", ".join((str(x) for x in self.PROJ))
-            raise ValueError(f"project '{project}' not in ({valid_projects})")
-        self.project = project
-        self.package = self.PROJ[project]
-        self.data_basedir = Path.home() / f".{self.project}"
-        try:
-            self.data_basedir.mkdir(parents=True, exist_ok=True)
-        except (FileNotFoundError, OSError) as err:
-            _log.error(f"error creating project data directory '{self.data_basedir}'")
-            raise
-        _log.info(f"Deployment: project={self.project} package={self.package} data_basedir={self.data_basedir}")
 
 
 class FlowsheetInfo(BaseModel):
